Split patcher as preparation for ottype compatiblity
